+++ core/src/main/scala/org/http4s/UrlForm.scala
@@ -81,6 +81,9 @@ object UrlForm {
   def apply(values: (String, String)*): UrlForm =
     values.foldLeft(empty)(_ + _)
 
+  def fromSeq(values: Seq[(String, String)]): UrlForm =
+    apply(values: _*)
+
   implicit def entityEncoder(implicit charset: Charset = DefaultCharset): EntityEncoder[UrlForm] =
     EntityEncoder.stringEncoder(charset)
       .contramap[UrlForm](encodeString(charset))